+++ tools/dependency/release_dates.py
@@ -16,6 +16,7 @@ import sys
 
 import github
 
+import exports
 import utils
