+++ tools/dependency/utils.py
@@ -5,26 +5,12 @@ from collections import namedtuple
 from importlib.util import spec_from_loader, module_from_spec
 from importlib.machinery import SourceFileLoader
 
-
-# Shared Starlark/Python files must have a .bzl suffix for Starlark import, so
-# we are forced to do this workaround.
-def LoadModule(name, path):
-  spec = spec_from_loader(name, SourceFileLoader(name, path))
-  module = module_from_spec(spec)
-  spec.loader.exec_module(module)
-  return module
+from exports import (
+  api_repository_locations, envoy_repository_locations, repository_locations_utils)
 
 
 # All repository location metadata in the Envoy repository.
 def RepositoryLocations():
-  api_path = os.getenv("API_PATH", "external/envoy_api_canonical")
-
-  envoy_repository_locations = LoadModule('envoy_repository_locations',
-                                          'bazel/repository_locations.bzl')
-  repository_locations_utils = LoadModule(
-      'repository_locations_utils', os.path.join(api_path, 'bazel/repository_locations_utils.bzl'))
-  api_repository_locations = LoadModule('api_repository_locations',
-                                        os.path.join(api_path, 'bazel/repository_locations.bzl'))
   spec_loader = repository_locations_utils.load_repository_locations_spec
   locations = spec_loader(envoy_repository_locations.REPOSITORY_LOCATIONS_SPEC)
   locations.update(spec_loader(api_repository_locations.REPOSITORY_LOCATIONS_SPEC))
